Green the P0 budget/guest/registry lint and tests
Final lint + test fixes for the P0 backend batch (tiered pricing, guest
limits, xAI registry row, budget soft-cap):

- usage.py: drop the redundant `int(...)` around `round()` in the soft-cap
  warning helper (RUF046) — `round(float)` already returns an int.
- test_bootstrap.py: the bootstrap provider-options set now includes the new
  pending `xai` route (T07).
- test_guest_limits.py: the substitution field is omitted from the wire when
  there is none, so assert via `.get(...)`; drop the redundant module-level
  asyncio mark (asyncio_mode=auto) that warned on the sync unit test.

+++ api/app/db/repositories/usage.py
@@ -108,7 +108,7 @@ def _budget_warning(effective_quota_usd: float, spend_usd: float) -> ErrorEnvelo
     if effective_quota_usd <= 0:
         return None
     ratio = spend_usd / effective_quota_usd
-    percent = int(round(ratio * 100))
+    percent = round(ratio * 100)
     if ratio >= 1.0:
         return platform_budget_soft_cap_envelope(percent=percent)
     threshold = get_settings().budget_warning_threshold_pct
